Оптимизация компонента main-page-sort-offers
