+++ project/package.json
@@ -28,7 +28,8 @@
       "htmlacademy/react-typescript"
     ],
     "rules": {
-      "no-console": "warn"
+      "no-console": "warn",
+      "comma-dangle": "warn"
     }
   },
   "browserslist": {
